confirmed to build on gcc and clang.

+++ ADAPT/FMP/Detail/FMPRe2c_impl.h
@@ -223,7 +223,7 @@ yy31:
 			}
 yy32:
 			{
-			yylval.Reset(stoll(this->Text()));
+			yylval.Reset((int64_t)stoll(this->Text()));
 			return FMP_TOKEN_INT;
 		}
 yy33:
